Adding a few more packages that were not in the manifest and moving some tests around.
